refactor(parser): split enhanced_statistics_parser by parse concern (#1125)
Behavior-preserving module split of the ~3,059-line monolith into
enhanced_statistics_parser/{mod,header,encoding_stats,marshal_type} plus a
serialization_header/{mod,sequential,schema} subtree. mod.rs keeps the public
entry points and re-exports the public surface unchanged; the parent
`pub use enhanced_statistics_parser::*` is untouched.

Concerns:
- header.rs        - 32-byte file header + TOC component-offset lookup
- encoding_stats.rs- EncodingStats (min ts/deletion-time/TTL) decode
- marshal_type.rs  - marshal-type -> CQL conversion + ColumnInfo builders
- serialization_header/ - schema parsing (dispatcher + offset-anchored decoders)
- mod.rs           - parse_nb_format_statistics_data / parse_enhanced_statistics_file
                     / parse_statistics_with_fallback orchestration

All 16 inline tests relocated next to their code (count preserved). No
semantic changes; nom parsers moved intact.
